+++ src/mod.ts
@@ -6,7 +6,7 @@ export { reduce } from "./fold.ts";
 export { count, cycle, repeat } from "./infinite.ts";
 export { map } from "./map.ts";
 export { Range, range } from "./range.ts";
-export { take, takeWhile } from "./take.ts";
+export { take, takeEnd, takeWhile } from "./take.ts";
 export { tee } from "./tee.ts";
 export { assertStreams, assertStreamStartsWith } from "./testing.ts";
 export { groupBy, unique } from "./unique.ts";